+++ src/app/[seasonSlug]/matches/[matchId]/page.tsx
@@ -27,7 +27,8 @@ import { getDemoDetail } from "@/actions/demo-detail";
 import { getTeamStyleProfile, getTeamHalfSideStats } from "@/actions/team-demo-stats";
 import { TeamStyleCompare } from "@/components/matches/TeamStyleCompare";
 import { playerRatings } from "@/db/schema";
-import { RadarChart, PRISM_AXES, buildPrismScores, averagePrismScores, type PrismScores } from "@/components/matches/RadarChart";
+import { RadarChart } from "@/components/matches/RadarChart";
+import { PRISM_AXES, buildPrismScores, averagePrismScores, type PrismScores } from "@/lib/stats/prism";
 import { DemoPlayerStatsTable } from "@/components/matches/DemoPlayerStatsTable";
 import { PlayerKillHeatmap } from "@/components/matches/PlayerKillHeatmap";
 import { PlayerWeaponBreakdown } from "@/components/matches/PlayerWeaponBreakdown";
