+++ src/app/players/[userId]/page.tsx
@@ -17,7 +17,8 @@ import { getSeasonHexagonScores } from "@/actions/hexagon";
 import type { HexagonScores } from "@/lib/utils/hexagon";
 import { getOcrAveragesBySeason } from "@/lib/stats";
 import { PlayerRadarChart } from "@/components/matches/PlayerRadarChart";
-import { RadarChart, PRISM_AXES, buildPrismScores, type PrismScores } from "@/components/matches/RadarChart";
+import { RadarChart } from "@/components/matches/RadarChart";
+import { PRISM_AXES, buildPrismScores, type PrismScores } from "@/lib/stats/prism";
 import { getPlayerDemoStats } from "@/actions/player-demo-stats";
 import { getSeasonWeaponStats, type PlayerWeaponStats } from "@/actions/season-demo-stats";
 import { PlayerDemoCard } from "@/components/players/PlayerDemoCard";
